Replace hatch-pet skill with v2 workflow
